extract compositions-overview ui into a dedicated component
