+++ scopes/compositions/compositions/index.ts
@@ -6,6 +6,5 @@ export type { CompositionsUI, CompositionsMenuSlot, EmptyStateSlot } from './com
 export { Composition, CompositionProps } from './composition';
 export { CompositionContent, CompositionContentProps } from './compositions';
 export type { CompositionsPreview } from './compositions.preview.runtime';
-export { CompositionsOverview } from './ui/compositions-overview';
 export { ComponentComposition } from './ui';
 export default CompositionsAspect;
