--- scopes/compositions/compositions/ui/compositions-overview/index.ts
@@ -1,1 +1,0 @@
-export * from './compositions-overview';
