+++ scopes/compositions/ui/compositions-overview/index.ts
@@ -1,0 +1,1 @@
+export {CompositionsOverview} from './compositions-overview';
