+++ scopes/react/ui/docs/compositions-carousel/compositions-carousel.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ApplyProviders } from '@teambit/react.ui.docs.apply-providers';
-import { CompositionsOverview } from '@teambit/compositions';
+import { CompositionsOverview } from '@teambit/compositions.ui.compositions-overview';
 import { LinkedHeading } from '@teambit/documenter.ui.linked-heading';
 import { Section, SectionProps } from '@teambit/documenter.ui.section';
 import { RenderingContext } from '@teambit/preview';
